refactor: rename stage entry points to runner.py for consistency with KGC
- corpus/update_bioc.py → corpus/runner.py
- filtering/run_filtering.py → filtering/runner.py
- extraction/run_extraction.py → extraction/runner.py
- Update all imports and test patches
